requirements.txt
python-telegram-bot
feedparser
schedule
deep-translator